+++ data/src/main/java/com/daejol/data/di/AppModule.kt
@@ -1,18 +1,16 @@
 package com.daejol.data.di
 
-import com.daejol.data.catdata.api.CatBreedsApi
 import com.daejol.data.ApiConst
+import com.daejol.data.catdata.api.CatBreedsApi
 import com.daejol.data.catdata.api.CatImagesApi
-import com.daejol.data.dogdata.api.DogImagesApi
 import com.daejol.data.dogdata.api.DogBreedsApi
+import com.daejol.data.dogdata.api.DogImagesApi
 import com.daejol.data.ranking.api.RankingApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Qualifier
@@ -103,47 +101,4 @@ object AppModule {
     fun provideCatRankingApi(@CatType retrofit: Retrofit): RankingApi {
         return retrofit.create(RankingApi::class.java)
     }
-
-    class CatRetrofitInterceptor : Interceptor {
-        override fun intercept(chain: Interceptor.Chain): Response {
-            val request = chain.request()
-            val params = mapOf<String, String>("api_key" to ApiConst.CAT_API_KEY)
-
-            val uriBuilder = request.url.newBuilder()
-            params.forEach { (key, value) ->
-                uriBuilder.addQueryParameter(key, value)
-            }
-
-            println("[keykat] $uriBuilder")
-
-            return chain.proceed(
-                request.newBuilder()
-                    .url(uriBuilder.build())
-                    .method(request.method, request.body)
-                    .build()
-            )
-        }
-    }
-
-    class DogRetrofitInterceptor : Interceptor {
-        override fun intercept(chain: Interceptor.Chain): Response {
-            val request = chain.request()
-            val params = mapOf<String, String>("api_key" to ApiConst.DOG_API_KEY)
-
-            val uriBuilder = request.url.newBuilder()
-            params.forEach { (key, value) ->
-                uriBuilder.addQueryParameter(key, value)
-            }
-
-            println("[keykat] $uriBuilder")
-
-            return chain.proceed(
-                request.newBuilder()
-                    .url(uriBuilder.build())
-                    .method(request.method, request.body)
-                    .build()
-            )
-        }
-    }
 }
-
